# Conflicts:
#	src/gameapi/ai/aisys/aiscript.cpp
#	src/gameapi/ai/aisys/aisys.cpp
#	src/gameapi/ai/aisys/aisys.h
#	src/gameapi/ai/aisys/gameai_actions.cpp
#	src/legoapi/items/base/apiobject.h

+++ src/legoapi/misc/utilities.cpp
@@ -28,8 +28,8 @@ void HashString(unsigned char *) {
 #ifdef _WIN32
 // MinGW lacks the POSIX srand48 family; nothing in the tree reads drand48
 // state, so seed the C RNG instead.
-static void srand48(long seed) {
-    srand(static_cast<unsigned int>(seed));
+static void srand48(i32 seed) {
+    srand(static_cast<u32>(seed));
 }
 #endif
 
@@ -167,7 +167,8 @@ bool LineIntersectCircle(NUVEC *origin, NUVEC *direction, NUVEC *center, f32 rad
     f32 x = center->x - origin->x;
     f32 z = center->z - origin->z;
     f32 projection = direction->x * x + direction->z * z;
-    if (projection >= 0.0f) return x * x + z * z - projection * projection <= radius_squared;
+    if (projection >= 0.0f)
+        return x * x + z * z - projection * projection <= radius_squared;
     return false;
 }
 
@@ -176,10 +177,13 @@ i32 LineIntersectSphere(NUVEC *origin, NUVEC *direction, NUVEC *center, f32 radi
     f32 y = center->y - origin->y;
     f32 z = center->z - origin->z;
     f32 projection = direction->x * x + direction->y * y + direction->z * z;
-    if (projection < 0.0f) return 0;
+    if (projection < 0.0f)
+        return 0;
     f32 distance = x * x + y * y + z * z - projection * projection;
-    if (radius_squared < distance) return 0;
-    if (distance_squared != NULL) *distance_squared = distance;
+    if (radius_squared < distance)
+        return 0;
+    if (distance_squared != NULL)
+        *distance_squared = distance;
     return 1;
 }
 
@@ -202,7 +206,8 @@ bool SphereSphereOverlap(NUVEC *a, f32 radius_a, NUVEC *b, f32 radius_b) {
 
 void CalcAveragePosAndRad(GIZBUILDIT_s &buildit, VuVec &position, float &radius, bool include_built) {
     u32 first = 0;
-    if (!include_built) first = buildit.built_object_count;
+    if (!include_built)
+        first = buildit.built_object_count;
     position = VuVec(0, 0, 0, 1);
     radius = 0.0f;
     i32 count = 0;
@@ -227,12 +232,15 @@ void CalcAveragePosAndRad(GIZBUILDIT_s &buildit, VuVec &position, float &radius,
             const f32 x = position.x - matrix->m30;
             const f32 z = position.z - matrix->m32;
             const f32 distance = x * x + 0.0f + z * z;
-            if (maximum <= distance) maximum = distance;
+            if (maximum <= distance)
+                maximum = distance;
         }
     }
-    if (maximum > 0.0f) maximum = NuFsqrt(maximum);
+    if (maximum > 0.0f)
+        maximum = NuFsqrt(maximum);
     radius = maximum;
-    if (buildit.radius_scale > 0.0f) radius = maximum * buildit.radius_scale;
+    if (buildit.radius_scale > 0.0f)
+        radius = maximum * buildit.radius_scale;
 }
 
 void LineToPlaneIntersecion(VuVec &, VuVec &, VuVec &, VuVec *) {
